moved fixtureref, adjusted tests
