--- pytest_docker_tools/wrappers/fixture_ref.py
@@ -1,10 +1,0 @@
-class FixtureRef:
-    def __init__(self, name):
-        self._name = name
-
-    @property
-    def name(self):
-        return self._name
-
-
-fixtureref = FixtureRef
